refactor(component-parser): split into smaller modules
