+++ src/parser/type-resolution.ts
@@ -1,0 +1,282 @@
+import type { LocalTypeDeclaration, ModernRunesTypeNode, ParsedComponentTypeScriptMetadata } from "../ComponentParser";
+import type { ParserContext } from "./context";
+import { sourceAtPos } from "./source-position";
+
+/**
+ * Looks up the modern-runes `$props()` declaration metadata recorded for the
+ * declarator starting at `declaratorStart`, if any.
+ */
+export function getRunesPropsDeclarationMetadata(ctx: ParserContext, declaratorStart: number | undefined) {
+  if (declaratorStart === undefined) return undefined;
+  return ctx.runesPropsDeclarationMetadataByDeclaratorStart.get(declaratorStart);
+}
+
+/**
+ * Looks up the per-prop type metadata (optionality, source range, type text)
+ * for `propName` within the `$props()` declaration starting at `declaratorStart`.
+ */
+export function getRunesPropTypeMetadata(ctx: ParserContext, declaratorStart: number | undefined, propName: string) {
+  return getRunesPropsDeclarationMetadata(ctx, declaratorStart)?.props.get(propName);
+}
+
+/** Rightmost name from a type reference (`Foo` or `NS.Foo`). */
+export function getTypeReferenceName(typeName: unknown): string | undefined {
+  if (!typeName || typeof typeName !== "object" || !("type" in typeName)) return undefined;
+
+  if (typeName.type === "Identifier" && "name" in typeName && typeof typeName.name === "string") {
+    return typeName.name;
+  }
+
+  if (
+    typeName.type === "TSQualifiedName" &&
+    "right" in typeName &&
+    typeName.right &&
+    typeof typeName.right === "object" &&
+    "name" in typeName.right &&
+    typeof typeName.right.name === "string"
+  ) {
+    return typeName.right.name;
+  }
+
+  return undefined;
+}
+
+/** Leftmost name from a type reference, for import/local lookup. */
+export function getTypeDependencyName(typeName: unknown): string | undefined {
+  if (!typeName || typeof typeName !== "object" || !("type" in typeName)) return undefined;
+
+  if (typeName.type === "Identifier" && "name" in typeName && typeof typeName.name === "string") {
+    return typeName.name;
+  }
+
+  if (typeName.type === "TSQualifiedName" && "left" in typeName && typeName.left && typeof typeName.left === "object") {
+    return getTypeDependencyName(typeName.left);
+  }
+
+  return undefined;
+}
+
+/** Returns the trimmed source text spanned by a type annotation node, if positions are available. */
+export function getTypeAnnotationText(
+  ctx: ParserContext,
+  typeAnnotation: { start?: number; end?: number } | undefined,
+) {
+  const start = typeAnnotation?.start;
+  const end = typeAnnotation?.end;
+  if (start === undefined || end === undefined) return undefined;
+  return sourceAtPos(ctx, start + 1, end)?.trim();
+}
+
+/** Collect imported and local type names referenced by a type AST node. */
+export function collectReferencedTypeDependencies(
+  ctx: ParserContext,
+  typeNode: ModernRunesTypeNode | undefined,
+  referencedImportedTypes: Set<string>,
+  referencedLocalTypes: Set<string>,
+  visitedLocalTypes: Set<string> = new Set(),
+) {
+  if (!typeNode?.type) return;
+
+  switch (typeNode.type) {
+    case "TSInterfaceDeclaration":
+      collectReferencedTypeDependencies(
+        ctx,
+        { type: "TSTypeLiteral", members: typeNode.body?.body },
+        referencedImportedTypes,
+        referencedLocalTypes,
+        visitedLocalTypes,
+      );
+      return;
+    case "TSTypeAliasDeclaration":
+    case "TSParenthesizedType":
+    case "TSTypeAnnotation":
+      collectReferencedTypeDependencies(
+        ctx,
+        typeNode.typeAnnotation,
+        referencedImportedTypes,
+        referencedLocalTypes,
+        visitedLocalTypes,
+      );
+      return;
+    case "TSIntersectionType":
+    case "TSUnionType":
+      for (const nestedType of typeNode.types ?? []) {
+        collectReferencedTypeDependencies(
+          ctx,
+          nestedType,
+          referencedImportedTypes,
+          referencedLocalTypes,
+          visitedLocalTypes,
+        );
+      }
+      return;
+    case "TSTypeLiteral":
+      for (const member of typeNode.members ?? []) {
+        if (member?.type !== "TSPropertySignature") continue;
+        collectReferencedTypeDependencies(
+          ctx,
+          member.typeAnnotation?.typeAnnotation,
+          referencedImportedTypes,
+          referencedLocalTypes,
+          visitedLocalTypes,
+        );
+      }
+      return;
+    case "TSTypeReference": {
+      const dependencyName = getTypeDependencyName(typeNode.typeName);
+      if (dependencyName) {
+        if (ctx.typeImportBindingsByLocalName.has(dependencyName)) {
+          referencedImportedTypes.add(dependencyName);
+        }
+
+        const localDeclaration = ctx.localTypeDeclarationsByName.get(dependencyName);
+        if (localDeclaration && !visitedLocalTypes.has(dependencyName)) {
+          referencedLocalTypes.add(dependencyName);
+          visitedLocalTypes.add(dependencyName);
+          collectReferencedTypeDependencies(
+            ctx,
+            localDeclaration.node,
+            referencedImportedTypes,
+            referencedLocalTypes,
+            visitedLocalTypes,
+          );
+          visitedLocalTypes.delete(dependencyName);
+        }
+      }
+
+      if ("typeParameters" in typeNode && typeNode.typeParameters && typeof typeNode.typeParameters === "object") {
+        const paramsNode = typeNode.typeParameters as { params?: ModernRunesTypeNode[] };
+        for (const param of paramsNode.params ?? []) {
+          collectReferencedTypeDependencies(
+            ctx,
+            param,
+            referencedImportedTypes,
+            referencedLocalTypes,
+            visitedLocalTypes,
+          );
+        }
+      }
+      return;
+    }
+    case "TSArrayType":
+    case "TSRestType":
+    case "TSOptionalType":
+    case "TSIndexedAccessType":
+    case "TSTypeOperator":
+    case "TSExpressionWithTypeArguments":
+    case "TSTupleType":
+    case "TSConditionalType":
+    case "TSInferType":
+    case "TSMappedType":
+    case "TSFunctionType":
+    case "TSConstructorType":
+    case "TSTypeQuery":
+    case "TSImportType":
+    case "TSLiteralType":
+    case "TSTypePredicate":
+    case "TSNamedTupleMember":
+      break;
+    default:
+      break;
+  }
+
+  for (const value of Object.values(typeNode)) {
+    if (!value || typeof value !== "object") continue;
+
+    if (Array.isArray(value)) {
+      for (const item of value) {
+        if (!item || typeof item !== "object" || !("type" in item)) continue;
+        collectReferencedTypeDependencies(
+          ctx,
+          item as ModernRunesTypeNode,
+          referencedImportedTypes,
+          referencedLocalTypes,
+          visitedLocalTypes,
+        );
+      }
+      continue;
+    }
+
+    if ("type" in value) {
+      collectReferencedTypeDependencies(
+        ctx,
+        value as ModernRunesTypeNode,
+        referencedImportedTypes,
+        referencedLocalTypes,
+        visitedLocalTypes,
+      );
+    }
+  }
+}
+
+/** Emit `import type` statements for referenced type names, grouped by module. */
+export function buildTypeImportStatements(ctx: ParserContext, referencedImportedTypes: Set<string>) {
+  const groupedImports = new Map<
+    string,
+    { default: string[]; named: Array<{ imported?: string; local: string }>; namespace: string[] }
+  >();
+
+  for (const importedType of Array.from(referencedImportedTypes).sort()) {
+    const binding = ctx.typeImportBindingsByLocalName.get(importedType);
+    if (!binding) continue;
+
+    const group = groupedImports.get(binding.source) ?? { default: [], named: [], namespace: [] };
+
+    if (binding.specifierType === "default") {
+      group.default.push(binding.localName);
+    } else if (binding.specifierType === "namespace") {
+      group.namespace.push(binding.localName);
+    } else {
+      group.named.push({
+        imported: binding.importedName,
+        local: binding.localName,
+      });
+    }
+
+    groupedImports.set(binding.source, group);
+  }
+
+  return Array.from(groupedImports.entries())
+    .sort(([sourceA], [sourceB]) => sourceA.localeCompare(sourceB))
+    .map(([source, group]) => {
+      if (group.namespace.length > 0) {
+        return group.namespace.map((localName) => `import type * as ${localName} from "${source}";`).join("\n");
+      }
+
+      const namedParts = group.named.map(({ imported, local }) => {
+        if (!imported || imported === local) return local;
+        return `${imported} as ${local}`;
+      });
+      const defaultImport = group.default[0];
+      const namedImport = namedParts.length > 0 ? `{ ${namedParts.join(", ")} }` : "";
+
+      if (defaultImport && namedImport) {
+        return `import type ${defaultImport}, ${namedImport} from "${source}";`;
+      }
+      if (defaultImport) {
+        return `import type ${defaultImport} from "${source}";`;
+      }
+      return `import type ${namedImport} from "${source}";`;
+    });
+}
+
+/** `ParsedComponentTypeScriptMetadata` for a single typed `$props()` declaration, if any. */
+export function buildTypeScriptMetadata(ctx: ParserContext): ParsedComponentTypeScriptMetadata | undefined {
+  if (ctx.typedRunesPropsDeclarations.length !== 1) return undefined;
+
+  const [typedDeclaration] = ctx.typedRunesPropsDeclarations;
+  if (!typedDeclaration.canonicalType) return undefined;
+
+  const localTypeDeclarations = Array.from(typedDeclaration.referencedLocalTypes)
+    .map((typeName) => ctx.localTypeDeclarationsByName.get(typeName))
+    .filter((declaration): declaration is LocalTypeDeclaration => declaration !== undefined)
+    .sort((a, b) => a.start - b.start)
+    .map((declaration) => declaration.code);
+
+  return {
+    canonicalPropsType: typedDeclaration.canonicalType,
+    canonicalPropNames: Array.from(typedDeclaration.props.keys()).sort(),
+    localTypeDeclarations,
+    typeImportStatements: buildTypeImportStatements(ctx, typedDeclaration.referencedImportedTypes),
+  };
+}
